transpose test

+++ transpose.cu
@@ -9,7 +9,7 @@ __global__ void transpose(float *matrix_t, float *matrix){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.x;
 
-    matrix_t[i] = matrix[blockDim.x * (i % gridDim.x) + floor(i / gridDim.x)];
+    matrix_t[i] = matrix[(int) (blockDim.x * (i % gridDim.x) + floor(i / gridDim.x))];
 }
